# docs/openapi/api.v1.json
{
  "components" : {
    "parameters" : {
      "CorrelationId" : {
        "description" : "Optional trace ID; generated if absent and returned on every response",
        "in" : "header",
        "name" : "X-Correlation-Id",
        "required" : false,
        "schema" : {
          "example" : "demo-123",
          "type" : "string"
        }
      }
    },
    "schemas" : {
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "correlationId" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "violations" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : "array"
          }
        }
      },
...
      "FieldViolation" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "JobView" : {
        "properties" : {
          "attempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "claimId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "correlationId" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "dueAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lastError" : {
            "type" : "string"
          },
          "maxAttempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING", "RUNNING", "DONE", "FAILED", "CANCELLED" ],
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageResponseJobView" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/JobView"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login (valid 15 minutes)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Guidewire-style claims handling: FNOL, exposures, reserves, payments, approvals, SIU and AI document assessment.\n\nAPI version **v1** (contract 1.3.0). Breaking changes get `/api/v2`; additions bump the minor version.",
    "title" : "AI Claims Platform API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/ops/jobs" : {
      "get" : {
        "description" : "Default: FAILED",
        "operationId" : "listJobs",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "default" : "FAILED",
            "enum" : [ "PENDING", "RUNNING", "DONE", "FAILED", "CANCELLED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseJobView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Jobs by status, most recently changed first",
        "tags" : [ "Operations" ]
      }
    },
    "/api/v1/ops/jobs/{id}/retry" : {
      "post" : {
        "operationId" : "retryJob",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Retry a FAILED job now, with a fresh set of attempts",
        "tags" : [ "Operations" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8081"
  } ],
...
}